Separating pack-onboarder into its own script
